Modernize 9ns networking function declarations
The naming collision between Plan 9 and BSD sockets needs to go.  If you
write a sockets program, but link in iplib, you'll get Plan 9's listen()
instead of the listen() in glibc.

When searching for a name, Ron came up with listen9 and accept9.  I love
it.  Hypothetical conversation:
	"Is that the 9th version of the accept call?"
	Nope.
	"Is it because it has 9 arguments?"
	Nope.
	"Is it because you belong in the 9th layer of hell?"
	Probably.

While we're changing the functions, I also added flags variables for
dial, announce, and listen.  I have uses for them later.