Rename event queue functions [2/2]
The common event queue is one with a mailbox, even if it is bigger.  So
the slim (no mailbox) is the special case.  Likewise, a raw eventq has a
mailbox, but the mailbox has not been initialized yet (used for UCQs and
later for CEQs).  And finally, the vcpd option remains a special case
(it's slim btw).

Nasty Awk for the function prototypes:

{
	gsub(/get_event_q_vcpd/, "get_eventq_vcpd")
	gsub(/get_event_q/, "get_eventq_slim")
	gsub(/get_big_event_q_raw/, "get_eventq_raw")
	gsub(/get_big_event_q/, "get_eventq")
	gsub(/put_event_q_vcpd/, "put_eventq_vcpd")
	gsub(/put_event_q/, "put_eventq_slim")
	gsub(/put_big_event_q_raw/, "put_eventq_raw")
	gsub(/put_big_event_q/, "put_eventq")
	print
}

And spatch for the function calls.  I couldn't get spatch to change the
prototypes.

@@
@@
-get_event_q
+get_eventq_slim
 (...)

@@
@@
-get_big_event_q
+get_eventq
 (...)

@@
@@
-get_big_event_q_raw
+get_eventq_raw
 (...)

@@
@@
-get_event_q_vcpd
+get_eventq_vcpd
 (...)

@@
@@
-put_event_q
+put_eventq_slim
 (...)

@@
@@
-put_big_event_q
+put_eventq
 (...)

@@
@@
-put_big_event_q_raw
+put_eventq_raw
 (...)

@@
@@
-put_event_q_vcpd
+put_eventq_vcpd
 (...)