Make O_RDONLY non-zero (XCC) [1/7]
O_RDONLY had the value of 0, which caused two problems.  A minor one was
that it was hard to check for readability or read-only-ness.  Another
was that it was impossible to have an FD that had no permissions.  You
were at least readable.  This was blocking some of our future plans.

As far as whether or not O_RDONLY must be zero:

http://pubs.opengroup.org/onlinepubs/009695399/functions/open.html

	In historical implementations the value of O_RDONLY is zero.
	Because of that, it is not possible to detect the presence of
	O_RDONLY and another option. Future implementations should
	encode O_RDONLY and O_WRONLY as bit flags so that:

	O_RDONLY | O_WRONLY == O_RDWR

So it looks like we're on solid ground.  Internally, we can refer to
O_RDONLY as O_READ too, since it's a little weird to say RDONLY |
WRONLY, since then they aren't "only."

This compiles, but you'll have runtime issues with the kernel.  The
following patches in the 5-patch series will fix those issues.

Rebuild the world.